Corte dos testes por assunto, com cada arquivo abaixo do teto de 350 linhas
O corte pedido eram dois arquivos; ficaram quatro, e o eixo é o mesmo que o
Manager pediu — um assunto por arquivo. Dois deles não cabiam:

  backfill (268)        §5 backfill inicial + guardas de escrita da projeção
  materializacao (255)  §4.1.1/§4.1.2 — os seis bloqueadores do dono
  projecao (269)        §4/§6 — projeção pura, ordem de eventos, varredura
  admin (144)           §15 — reparo manual e isolamento por usuário

Com o corte em dois, o de projeção passaria de 600 linhas: o dobro do teto que
motivou o corte.

`tests/_billing_grants_helpers.py` (sem prefixo `test_`, não coletado) guarda
`conta`/`ler`/`grant`/os construtores de evento, para os quatro arquivos não
carregarem quatro cópias do mesmo helper.

Ele também ganhou `garantir_system_event_logs`: essa tabela nasce no
`ensure_admin_tables` do startup, não no `init_db`, então num banco de teste ela
não existe e TODA dedup via `recent_event_exists` virava no-op silencioso — foi
por isso que o teste de reentrega acusou 3 e-mails onde devia haver 1.

Sete mutações refeitas depois dos dois cortes; nenhum teste virou no-op.
